Rewrote the csvHeader in State to use a vector for easier modifying and improved readability.
Also added a trailing comma to all of the sensors' data strings that then gets removed at the end. this way multiple sensors dont try and write data to the same column. A better way to do all of this would be to have a list of Sensors and iterate though the sensors to get their data strings/headers. Will implement this later.