Update & finished mag
Have fully implemented algo from Github. Process is that Mag independently takes in 30 seconds of dynamic readings at 100 Hz, stores them in a buffer, and then extracts that buffer data for the algo. The buffer after calibration is removed. It then uses those calibration matrices to perform calculations on the mag readings as it flies (easy matrix multiplication)