Fixed concerns from Drew
Hey Jack, just a few things I noticed about your PCB
First, somehow this resistor got a diode's footprint. Not the biggest issue but its kinda funny (pic 1)
Generally, it's better to have a continuous ground plane under everything and only run traces on one layer. Idk if it will have any real effect on what we're doing here, so if you don't have time or whatever you can just leave it, but for future projects, it's something to keep in mind. ground planes help with EMI and just generally help things perform better.
You should connect all of the GND pins on the Teensy to GND (pic 2)
This is a bit weird but its fine lol (pic 3)
Finally, you should add some silkscreen with your name on it and maybe label which connector is which for easier use.
Otherwise it all looks good